medivault-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
web3==6.11.3
ipfshttpclient==0.8.0a2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
cryptography>=41.0.0
